Rewrite `swap` to handle multiple inputs and be stateless
In the process of this, we define the roles of Alice and Bob as:

Alice: Is always the taker.
Bob: Is always the maker.

As the maker, Bob is the one to create the transaction and determine
an appropriate fee rate. The party contributing the native asset to
the transaction will pay for the fee.

Fees on Liquid are extremely low at the moment because there are basically
no transactions.
Once this situation changes, Bob should make the following considerations:

- If he is contributing the L-BTC, he is paying for the fee.
In this situation, he might charge Alice for her share by lowering the rate.

- If Alice is contributing the L-BTC, she is paying for the fee.
In this situation, Bob might reduce the amount of whatever asset he is selling
to Alice. This will allow him to take some of Alice's L-BTC to pay for the fee
without having to adapt the rate.

In addition to these changes, we also move from two addresses per actor to a
single one. This doesn't have major privacy implications but simplifies the
implementation.