Replace `Lazy` statics with constants
We add a buildscript that reads the env variables, verifies them and
writes them to a Rust file as constants which we can include at
build time.

This allows us to verify all the configuration parameters at build time
instead of being surprised at runtime with an error like
"invalid elements chain".