Make it clear that `make_create_swap_payload` is creating a _sell_ swap
This function assumes that the given BTC amount is to be sold rather
than bought. Renaming the function makes this obvious.

The payload struct itself remains the same because it can be reused for
the buying usecase where it can express how much BTC the party is intending
to buy.