Refactor state handling within the app
The core motivation of this patch is to improve UX by displaying
loading spinners where appropriate and sequencing events in a way
that improves the UX (i.e. don't close drawers until async actions
are done etc).

We achieve this by pulling most of the state into App.tsx and make
the individual components stateless for anything that is not contained
within them. For example, the drawers on longer manage their own
open/close state, only the state of their inner form.

We also reorganise the code such that there is always just one button
displayed and we modify the click handler and label according to our
current state.

The work started in this patch is not fully complete but it sets the
ground for continuing in a certain direction. For example, we likely
want to merge all the drawers into one and only replace the content
because we only always display just one of them at a time. As such,
it would be easier to understand if that is reflected in the code.