Stop making too many evil clones of Bob
Bobtimus holds an instance of rng initialised at start-up. Every time
we receive a request, we clone our instance of Bobtimus and therefore
the instance of rng itself. Any state changes to the rng (which happen
during usage) thereby did not affect the original prototype of
Bobtimus. As a result, any future request started with exactly the
same rng state and therefore produced the same sequence of random
numbers.

This led to conflicting blinding factors between inputs and outputs.

Fixes #91.