Serve frontend under /app namespace
This applies to all assets apart from index.html.

Unfortunately, we need to add a hook to the app to redirect from
/app to / to make it work with the webpack-dev-server because
setting `homepage` in `package.json` makes `react-scripts` open
the browser on `http://localhost:3004/app`.

I hesitated to guard this piece of code with `if env == development`
because I think it might actually also be useful in production.
If someone for whatever reasons investigates our routes and finds
`/app`, we just send them back to the root.

Resolve #79.