Make the amounts less jumpy.
By clicking on the round circle in the middle alpha amount/asset and beta amount/asset are swapped. Due to different ask and bid prices the amounts would weirdly change and decrease on every swap.
Instead of complex calculation we simply fix the alpha amount and only swap the asset types.