Make CreateSwapPayload amount express Alice's amount
Instead of forcing Alice to always determine the amount of bitcoin
involved in the swap, it is more natural for her to send the amount
she is paying for with her inputs.

We replace `sell_quote` with `sell_base` to calculate how much bitcoin
Alice should get when she offers a certain amount of USDt.