Forward unmatched url to webapp.
So far, we forwarded any request at `/` to index.thml. However, we use client side routing in the webapp which allows us to have relative urls e.g. `/swapped/{txid}`. When triggered on the client (e.g. by clicking a button) this is not an issue as the react router takes care of everything. If however a request is forwarded to the server (bobtimus) an HTTP-403 is returned.
With this patch we redirect any un-matched url to the webapp and let the client side routing do the rest.