Fix repayLoan internal extension API
Bobtimus was using non-segwit change outputs which was causing the loan
txid to change after signing. This prevented the wallet extension from
finding the loan transaction on the blockchain, and it would have made
it impossible for the repayment to take place, since the txid of the
collateral input would have been wrong.

Additionally, the array of open loans was not bein updated after
repaying a loan because we were attempting to delete the entry
corresponding to the _repayment_ txid rather than the _loan_ txid.