Fix getOpenLoans internal extension API
We were only saving the txid as opposed to the `LoanDetails`. Now we
save state and `LoanDetails` together after extracting the loan, so
that when we get the state out to sign the loan we also have the
`LoanDetails` available and we can add them to the array of open loans
inlocal storage.