Fix SSE rate service.
The issue is some compression feature in webpack-dev server which does not let the SSE events through.
By telling the client to not apply any cache transformations, the message goes through.
The result is that events are not cached or buffered in anyway which could be a performance issue.
While this is only a dev issue for us it should have hardly any implications in production neither as our events are super small.