Docker compose based e2e test setup with config.
The docker-compose file is inspired by nigiri but we removed Bitcoin from the setup. This reduces the startup time and resource consumption.
Unfortunately it was not possible to use virtual volumes.
Hence it is required to clear the local blockchain data from time to time which can be found in : `nigiri/liquid-config/liquidregtest/*`

Signed-off-by: Philipp Hoenisch <philipp@hoenisch.at>