Fix tooltip display
For wrapping a custom component with a tooltip on has to implement forwardRef for the component. Since this is a bit cumbersome another workaround is to wrap the custom component with a Box.
From chakra-ui documentation: If the Tooltip is wrapping a functional component, ensure that the functional component accepts a ref using forwardRef.