Sort UTXOs to have more deterministic logs
This should help us to see more deterministic logs upon repay because the UTXOs that are skipped should always be in the same order.
An unordered vec does not cause (known) bugs, but the log output can be confusing. Since this sorting should not be expensive we put it in.