Lock selected utxos to not double select them.
This fixes an issue were bobtimus tries to double spend himself. If we do not lock the selected utxos bobtimus might just reselect them. This can cause a race condition if two competing transactions are trying to spend the same utxo.

Note: this can cause side effects were bobtimus runs out of money. If this happens, we need to manually mine a block.
Signed-off-by: Philipp Hoenisch <philipp@hoenisch.at>