Directly return errors from background script to in-page script
The listener in `browser.runtime.onMessage.addListener` can handle
promises. As such, we can directly _reject_ this promise if there
is any error in invoking the various wallet functions.

Such a rejected promise is then _caught_ inside the content script
and forwarded as an event to the in-page script. The communication
of the events between in-page script and content-script has been
reworked.

1. The new implementation contains less duplicated code due to the
use of a generic `invokeContentScript` function.

2. Every request is tagged with an ID which allows us to uniquely
identify, which particular request failed. This should help with
race conditions where the same event is fired multiple times.

Related: #196.