Ensure we don't have any floating promises
Floating promises make awaiting a function useless because the promise
is just executed in the background and the function returns immediately.
In our case, this led to the onSuccess() handler in the ConfirmSwap component
being called too early and hence the page refreshed at a point where the
old state (i.e. the swap to sign) was still present.

Fixes #244.