Re-architect message passing in extension
This patch completely re-architectes the message passing to cut down
on boilerplate code and make things more understandable.
Most importantly with this patch, we no longer treat the WASM language
boundary as a guidance on which functionality should be implemented where.

Fpr example, previously all event listeners where implemented in Typescript.
These event listeners would then forward to functions exposed from the Rust
code via `wasm_bindgen`. This indrection causes unnecessary boilerplate.

Now, we primarily differentiate between two APIs:

1. Functions injected into the web page via the in-page script
2. Privileged APIs used by the frontend of the extension

The former is implemented via message passing through the content script.
Some of the event handlers for these messages are implemented in Rust, some
are implemented in Typescript, depending on what is more convenient.
Specifically, anything with user interaction is implemented in Typescript because
it allows us to temporarily store the resolve functions of the promise.

The privileged APIs are exposed as functions on the background-script's
`window` object. This window object can be accessed from any other privileged
script. Not only is this convenient, it also allows us to use the browser's
priviledge-enforcement code to ensure that a malicious web page cannot
call these APIs and they are truly internal to the extension.

Resolves #228.